+++ raqqoon-sidedish/src/components/TabUI/TabBody.jsx
@@ -1,6 +1,6 @@
-import { useState } from 'react';
 import styled from 'styled-components';
 import useFetch from '../../customHooks/useFetch';
+import useTab from '../../customHooks/useTab';
 import TabList from './TabList';
 import TabMain from './TabMain';
 
@@ -9,16 +9,6 @@ const TabBody = () => {
     `https://h3rb9c0ugl.execute-api.ap-northeast-2.amazonaws.com/develop/baminchan/best`,
     []
   );
-
-  const useTab = (tabIdx, tabItemList) => {
-    const [tabIndex, setTabIndex] = useState(tabIdx);
-
-    return {
-      tabItemList: tabItemList[tabIndex] || [],
-      tabIndex: tabIndex,
-      setTabIndex: setTabIndex,
-    };
-  };
 
   const { tabItemList, tabIndex, setTabIndex } = useTab(0, bestData);
